카테고리 하위 누르면 에러 터지는거 해결 + 이미지 파일 깨지는것해결까지

+++ src/main/java/ljh/commerce/commerce/domain/product/ProductRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ProductRepository extends JpaRepository<Product,Long> {
-    List<Product> findAllByCategoryId(int categoryId);
+    List<Product> findAllByCategoryId(long categoryId);
 }
